Move tests to Functional namespace
