+++ tests/Functional/Api/TripsApiTest.php
@@ -1,11 +1,11 @@
 <?php
 declare(strict_types=1);
 
-namespace Xvilo\OVpayApi\Tests\Api;
+namespace Xvilo\OVpayApi\Tests\Functional\Api;
 
 use Symfony\Component\HttpClient\Response\MockResponse;
 use Xvilo\OVpayApi\Authentication\HeaderMethod;
-use Xvilo\OVpayApi\Tests\TestCase;
+use Xvilo\OVpayApi\Tests\Functional\TestCase;
 
 final class TripsApiTest extends TestCase
 {
